Set travis-ci script cmd to rspec
Why you made the change:

I did this instead of the default `rake` that it does because I am not
getting code climate code coverage results. According to their website
http://docs.codeclimate.com/article/220-help-im-having-trouble-with-test-coverage
in their trouble shooting section it will have if launched from rake
because it spins up another process. Hopefully, this makes it work.