Make middleware thread safe for instance vars
Why you made the change:

I did this so that we could use instance variables safely with Puma or
other threaded rack app servers.

How the change addresses the need:

Having the `call` method first `dup` the instance of the middleware
creates a shallow copying (meaning all its instance variables are
coppied) of the middleware. This is important because next it calls
`_call` on the dup. This makes it so that any alteration of instance
variables that happen within a request (and a thread) or scoped to the
copy not the original instance which is referenced when creating new
threads.